# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bengal Creations Marketplace API",
    "description": "Complete REST API for a multi-vendor marketplace — vendors, products, categories, cart, orders, payments, wishlist, and addresses.",
    "version": "1.0.0"
  },
  "host": "api.bengalcreations.in",
  "basePath": "/",
  "schemes": ["https", "http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter your JWT token as: Bearer <token>"
    }
  },
...
  "paths": {
    "/api/contact": {
      "post": {
        "tags": ["Contact"],
        "summary": "Submit a contact form message",
        "parameters": [{
          "in": "body", "name": "body", "required": true,
          "schema": { "$ref": "#/definitions/ContactFormBody" }
        }],
        "responses": {
          "200": { "description": "Message received" },
          "500": { "description": "Server error" }
        }
      }
    },
...
    "/api/vendors": {
      "get": {
        "tags": ["Vendors"],
        "summary": "Get all vendors with pagination",
        "parameters": [
          { "in": "query", "name": "page",     "type": "integer", "default": 1 },
          { "in": "query", "name": "limit",    "type": "integer", "default": 10 },
          { "in": "query", "name": "verified", "type": "boolean" }
        ],
        "responses": {
          "200": { "description": "Paginated vendor list", "schema": { "$ref": "#/definitions/PaginatedVendors" } }
        }
      }
    },
...
    "/api/products": {
      "post": {
        "tags": ["Products"],
        "summary": "Create a new product",
        "parameters": [{
          "in": "body", "name": "body", "required": true,
          "schema": { "$ref": "#/definitions/ProductBody" }
        }],
        "responses": {
          "201": { "description": "Product created", "schema": { "$ref": "#/definitions/Product" } },
          "400": { "description": "Request body missing" }
        }
      },
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "responses": {
          "200": { "description": "All products" }
        }
      }
    },
...
    "/api/categories": {
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category (with optional image upload)",
        "consumes": ["multipart/form-data"],
        "parameters": [
          { "in": "formData", "name": "name",   "required": true, "type": "string" },
          { "in": "formData", "name": "parent", "type": "string" },
          { "in": "formData", "name": "image",  "type": "file" }
        ],
        "responses": {
          "201": { "description": "Category created", "schema": { "$ref": "#/definitions/Category" } }
        }
      },
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "responses": {
          "200": { "description": "All categories" }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create an order from the customer cart",
        "parameters": [{
          "in": "body", "name": "body", "required": true,
          "schema": { "$ref": "#/definitions/CreateOrderBody" }
        }],
        "responses": {
          "200": { "description": "Order created", "schema": { "$ref": "#/definitions/Order" } },
          "400": { "description": "Cart empty or item out of stock" },
          "404": { "description": "Address or product not found" }
        }
      },
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders",
        "responses": {
          "200": { "description": "All orders" }
        }
      }
    },
...
    "/api/addresses": {
      "post": {
        "tags": ["Addresses"],
        "summary": "Add a new address",
        "parameters": [{
          "in": "body", "name": "body", "required": true,
          "schema": { "$ref": "#/definitions/AddressBody" }
        }],
        "responses": {
          "200": { "description": "Address saved", "schema": { "$ref": "#/definitions/Address" } },
          "500": { "description": "Server error" }
        }
      }
    },
...
    "/api/wishlist": {
      "get": {
        "tags": ["Wishlist"],
        "summary": "Get the authenticated customer wishlist",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Wishlist with populated products" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Vendor": {
      "type": "object",
      "properties": {
        "vendorId":    { "type": "string", "example": "VEND-20241A2B3C" },
        "name":        { "type": "string" },
        "shopName":    { "type": "string" },
        "email":       { "type": "string" },
        "phone":       { "type": "string" },
        "address":     { "type": "string" },
        "description": { "type": "string" },
        "logo":        { "type": "string" },
        "banner":      { "type": "string" },
        "isVerified":  { "type": "boolean" },
        "documents": {
          "type": "object",
          "properties": {
            "tradeLicense": { "type": "string" },
            "aadhaarCard":  { "type": "string" },
            "panCard":      { "type": "string" }
          }
        },
        "createdAt": { "type": "string", "format": "date-time" }
      }
    },
...
    "PaginatedVendors": {
      "type": "object",
      "properties": {
        "success": { "type": "boolean" },
        "vendors": { "type": "array", "items": { "$ref": "#/definitions/Vendor" } },
        "pagination": {
          "type": "object",
          "properties": {
            "total": { "type": "integer" },
            "page":  { "type": "integer" },
            "limit": { "type": "integer" },
            "pages": { "type": "integer" }
          }
        }
      }
    },
    "ProductBody": {
      "type": "object",
      "required": ["name","price","stock","district","vendor"],
      "properties": {
        "name":          { "type": "string",  "example": "Handwoven Silk Saree" },
        "description":   { "type": "string" },
        "price":         { "type": "number",  "example": 4500 },
        "originalPrice": { "type": "number",  "example": 5500 },
        "stock":         { "type": "integer", "example": 20 },
        "district":      { "type": "string",  "example": "Varanasi" },
        "category":      { "type": "string",  "example": "64f1a2b3c4d5e6f7a8b9c0d2" },
        "vendor":        { "type": "string",  "example": "64f1a2b3c4d5e6f7a8b9c0d3" },
        "images":        { "type": "array",   "items": { "type": "string" } }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "_id":         { "type": "string" },
        "name":        { "type": "string" },
        "description": { "type": "string" },
        "price":       { "type": "number" },
        "orginalPrice":{ "type": "number" },
        "stock":       { "type": "integer" },
        "district":    { "type": "string" },
        "images":      { "type": "array", "items": { "type": "string" } },
        "isActive":    { "type": "boolean" },
        "category":    { "type": "object" },
        "vendor":      { "type": "object" },
        "createdAt":   { "type": "string", "format": "date-time" }
      }
    },
...
    "Category": {
      "type": "object",
      "properties": {
        "_id":      { "type": "string" },
        "name":     { "type": "string", "example": "Sarees" },
        "slug":     { "type": "string", "example": "sarees" },
        "image":    { "type": "string" },
        "parent":   { "type": "string", "nullable": true },
        "isActive": { "type": "boolean" }
      }
    },
    "CreateOrderBody": {
      "type": "object",
      "required": ["user_id","addressId","PaymentMethod"],
      "properties": {
        "user_id":       { "type": "string", "example": "64f1a2b3c4d5e6f7a8b9c0d1" },
        "addressId":     { "type": "string", "example": "64f1a2b3c4d5e6f7a8b9c0e1" },
        "PaymentMethod": { "type": "string", "enum": ["COD","Online"], "example": "Online" }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "_id":           { "type": "string" },
        "user":          { "type": "string" },
        "items":         { "type": "array" },
        "totalAmount":   { "type": "number" },
        "paymentMethod": { "type": "string", "enum": ["COD","Online"] },
        "paymentStatus": { "type": "string", "enum": ["Pending","Paid","Failed"] },
        "status":        { "type": "string", "enum": ["Pending","Processing","Shipped","Delivered","Cancelled"] },
        "createdAt":     { "type": "string", "format": "date-time" }
      }
    },
...
    "AddressBody": {
      "type": "object",
      "properties": {
        "customer":  { "type": "string" },
        "fullName":  { "type": "string", "example": "Jane Doe" },
        "phone":     { "type": "string", "example": "9876543210" },
        "pincode":   { "type": "string", "example": "110001" },
        "state":     { "type": "string", "example": "Delhi" },
        "city":      { "type": "string", "example": "New Delhi" },
        "area":      { "type": "string", "example": "Connaught Place" },
        "houseNo":   { "type": "string", "example": "Flat 101, Block A" },
        "landmark":  { "type": "string", "example": "Near Metro Station" }
      }
    },
    "Address": {
      "type": "object",
      "properties": {
        "_id":      { "type": "string" },
        "customer": { "type": "string" },
        "fullName": { "type": "string" },
        "phone":    { "type": "string" },
        "pincode":  { "type": "string" },
        "state":    { "type": "string" },
        "city":     { "type": "string" },
        "area":     { "type": "string" },
        "houseNo":  { "type": "string" },
        "landmark": { "type": "string" }
      }
    },
...
  }
}